// _src/tsBase/svc/_Var.d.ts
export default class _Var {
    static preZero(len: number, value: string | number): string;
    static emptyToValue(var1: any, value: any): any;
    static isEmpty(var1: any): boolean;
    static isStr(var1: any): boolean;
    static notEmpty(var1: any): boolean;
    static isPureData(value: any): boolean;
    static toBool(val: any): boolean;
}